+++ mirrulations-dashboard/src/mirrdash/static/index.js
