+++ mirrulations-dashboard/src/mirrdash/static/style.css
