+++ mirrulations-dashboard/src/mirrdash/templates/index.html
