+++ mirrulations-dashboard/tests/test_dashboard.py
@@ -1,6 +1,16 @@
 from pytest import fixture
-from c21server.dashboard.dashboard_server import create_server, get_jobs_stats
-from .test_utils import mock_flask_server
+from fakeredis import FakeRedis, FakeServer
+from mirrdash.dashboard_server import create_server, get_jobs_stats
+
+
+def mock_flask_server(create_server):
+    redis_server = FakeServer()
+    mock_db = FakeRedis(server=redis_server)
+    server = create_server(mock_db)
+    server.redis_server = redis_server
+    server.app.config['TESTING'] = True
+    server.client = server.app.test_client()
+    return server
 
 
 @fixture(name='mock_server')
